+++ components/SocialIconsRow/SocialIconsRow.tsx
@@ -1,4 +1,4 @@
-import { Link } from '../Link';
+import { Link } from '../Link/Link';
 import { Icon } from '../Icon/Icon';
 import styles from './SocialIconsRow.module.css';
