+++ pages/index.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { GetStaticProps } from 'next';
 import { Accent } from '../components/Accent/Accent';
 import { MainLayout } from '../components/Layout';
-import { Link } from '../components/Link';
+import { Link } from '../components/Link/Link';
 import { getAllPosts, Post } from '../utils/posts';
 import { Spacer } from '../components/Spacer';
 import { PostPreviewList } from '../components/PostPreviewList';
